Refactor for solve merge conflicts
