+++ src/test/java/faang/school/projectservice/service/project/ProjectValidatorTest.java
@@ -10,6 +10,7 @@ import faang.school.projectservice.model.ProjectVisibility;
 import faang.school.projectservice.model.Team;
 import faang.school.projectservice.model.TeamMember;
 import faang.school.projectservice.repository.ProjectRepository;
+import faang.school.projectservice.validator.project.ProjectValidator;
 import jakarta.persistence.EntityNotFoundException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -28,9 +29,9 @@ import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-public class ProjectServiceValidatorTest {
+public class ProjectValidatorTest {
     @InjectMocks
-    private ProjectServiceValidator validator;
+    private ProjectValidator validator;
     @Mock
     private ProjectRepository repository;
     @Mock
